pyproject.toml
[tool.poetry]
name = "sisyphus-server"
version = "0.9.0"
description = ""
authors = ["JamesTheBard <jamesthebard@gmail.com>"]
readme = "README.md"
packages = [{include = "sisyphus_server"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
flask-restx = "^1.1.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
pycouchdb = "^1.14.2"
python-box = "^7.1.1"
humanize = "^4.8.0"
pymongo = "^4.5.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
